refactor(input): move from experimental to internal
BREAKING CHANGE: Input has moved to src/__internal__/input/
