+++ src/__internal__/input/input-sizes.style.js
