+++ src/components/date-range/date-range.style.js
@@ -2,7 +2,7 @@ import styled from "styled-components";
 import { margin } from "styled-system";
 import StyledDateInput from "../date/date.style";
 import { StyledLabelContainer } from "../../__internal__/label/label.style.js";
-import { StyledInputPresentationContainer } from "../../__experimental__/components/input/input-presentation.style.js";
+import { StyledInputPresentationContainer } from "../../__internal__/input/input-presentation.style.js";
 import baseTheme from "../../style/themes/base";
 
 const StyledDateRange = styled.div`
