+++ src/components/pager/__internal__/pager-navigation.spec.js
@@ -3,8 +3,8 @@ import { shallow, mount } from "enzyme";
 import PagerNavigation from "./pager-navigation.component";
 import { StyledPagerLinkStyles, StyledPagerNavInner } from "../pager.style";
 import { assertStyleMatch } from "../../../__spec_helper__/test-utils";
-import StyledInputPresentation from "../../../__experimental__/components/input/input-presentation.style";
-import StyledInput from "../../../__experimental__/components/input/input.style";
+import StyledInputPresentation from "../../../__internal__/input/input-presentation.style";
+import StyledInput from "../../../__internal__/input/input.style";
 
 const pageSizeSelectionOptions = [
   { id: "10", name: 10 },
