+++ src/components/pager/pager.style.js
@@ -1,6 +1,6 @@
 import styled, { css } from "styled-components";
-import StyledInput from "../../__experimental__/components/input/input.style";
-import StyledInputPresentation from "../../__experimental__/components/input/input-presentation.style";
+import StyledInput from "../../__internal__/input/input.style";
+import StyledInputPresentation from "../../__internal__/input/input-presentation.style";
 import InputIconToggleStyle from "../../__internal__/input-icon-toggle/input-icon-toggle.style";
 import Select from "../select/simple-select/simple-select.component";
 import baseTheme from "../../style/themes/base";
