+++ src/components/select/multi-select/multi-select.style.js
@@ -3,8 +3,8 @@ import StyledPill from "../../pill/pill.style";
 import { baseTheme } from "../../../style/themes";
 import InputIconToggleStyle from "../../../__internal__/input-icon-toggle/input-icon-toggle.style";
 import StyledSelect from "../select.style";
-import InputPresentationStyle from "../../../__experimental__/components/input/input-presentation.style";
-import sizes from "../../../__experimental__/components/input/input-sizes.style";
+import InputPresentationStyle from "../../../__internal__/input/input-presentation.style";
+import sizes from "../../../__internal__/input/input-sizes.style";
 
 const StyledSelectPillContainer = styled.div`
   display: flex;
