+++ src/components/select/select.style.js
@@ -1,8 +1,8 @@
 import styled, { css } from "styled-components";
 import { margin } from "styled-system";
 
-import InputPresentationStyle from "../../__experimental__/components/input/input-presentation.style";
-import StyledInput from "../../__experimental__/components/input/input.style";
+import InputPresentationStyle from "../../__internal__/input/input-presentation.style";
+import StyledInput from "../../__internal__/input/input.style";
 import InputIconToggleStyle from "../../__internal__/input-icon-toggle/input-icon-toggle.style";
 import { baseTheme } from "../../style/themes";
