+++ src/components/table/table-cell/table-cell.spec.js
@@ -5,8 +5,8 @@ import { ThemeProvider } from "styled-components";
 import TableCell from ".";
 import { Table, TableRow } from "..";
 import StyledTableCell from "./table-cell.style";
-import StyledInputPresentation from "../../../__experimental__/components/input/input-presentation.style";
-import StyledInput from "../../../__experimental__/components/input/input.style";
+import StyledInputPresentation from "../../../__internal__/input/input-presentation.style";
+import StyledInput from "../../../__internal__/input/input.style";
 import { rootTagTest } from "../../../utils/helpers/tags/tags-specs";
 import BaseTheme from "../../../style/themes/base";
 import mintTheme from "../../../style/themes/mint";
