+++ src/components/textbox/textbox.spec.js
@@ -8,7 +8,7 @@ import {
   testStyledSystemMargin,
 } from "../../__spec_helper__/test-utils";
 import FormField from "../../__experimental__/components/form-field";
-import InputPresentation from "../../__experimental__/components/input/input-presentation.component";
+import InputPresentation from "../../__internal__/input/input-presentation.component";
 import StyledValidationIcon from "../../__internal__/validations/validation-icon.style";
 import StyledPrefix from "./__internal__/prefix.style";
 import Label from "../../__internal__/label";
